+++ x/observer/types/codec.go
@@ -11,10 +11,10 @@ func RegisterCodec(cdc *codec.LegacyAmino) {
 	cdc.RegisterConcrete(&MsgAddObserver{}, "observer/AddObserver", nil)
 	cdc.RegisterConcrete(&MsgUpdateChainParams{}, "observer/UpdateChainParams", nil)
 	cdc.RegisterConcrete(&MsgRemoveChainParams{}, "observer/RemoveChainParams", nil)
+	cdc.RegisterConcrete(&MsgVoteBlockHeader{}, "observer/VoteBlockHeader", nil)
 	cdc.RegisterConcrete(&MsgAddBlameVote{}, "observer/AddBlameVote", nil)
 	cdc.RegisterConcrete(&MsgUpdateCrosschainFlags{}, "observer/UpdateCrosschainFlags", nil)
 	cdc.RegisterConcrete(&MsgUpdateKeygen{}, "observer/UpdateKeygen", nil)
-	cdc.RegisterConcrete(&MsgAddBlockHeader{}, "observer/AddBlockHeader", nil)
 	cdc.RegisterConcrete(&MsgUpdateObserver{}, "observer/UpdateObserver", nil)
 	cdc.RegisterConcrete(&MsgResetChainNonces{}, "observer/ResetChainNonces", nil)
 	cdc.RegisterConcrete(&MsgVoteTSS{}, "observer/VoteTSS", nil)
@@ -28,7 +28,7 @@ func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
 		&MsgAddBlameVote{},
 		&MsgUpdateCrosschainFlags{},
 		&MsgUpdateKeygen{},
-		&MsgAddBlockHeader{},
+		&MsgVoteBlockHeader{},
 		&MsgUpdateObserver{},
 		&MsgResetChainNonces{},
 		&MsgVoteTSS{},